Island-Bali/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

celery==5.3.6
cryptography==41.0.5
Django==4.2.6
django-phonenumber-field==6.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
environ==1.0
loguru==0.7.2
pandas==2.1.4
phonenumbers==8.13.23
qrcode==7.4.2
requests==2.31.0
Telethon==1.33.1
yandex-music==2.1.1
gunicorn==20.0.4
psycopg2-binary==2.9.1
python-dotenv==1.0.0
django-environ==0.10.0
django-jazzmin==2.6.0
django-celery-beat==2.5.0
django-extensions==3.2.3
Pillow==10.1.0
redis
sqlalchemy
